requirements.txt
asgiref==3.7.2
charset-normalizer==2.0.9
certifi==2021.10.8
cachetools==5.3.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.9
django-allauth
asgiref==3.7.2
Django
defusedxml==0.7.1
django-browser-reload==1.12.0
django-allauth==0.61.0
django-nested-admin==4.0.2
django-quill-editor==0.1.40
dj-database-url
et-xmlfile==1.1.0
httplib2==0.22.0
gunicorn==20.1.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
gspread==6.0.2
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
pyasn1==0.5.1
pytz==2021.3
python-dateutil
python-decouple==3.5
pyasn1-modules==0.3.0
Pillow
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-monkey-business==1.0.0
python3-openid==3.2.0
requests==2.26.0
reportlab==4.1.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
sqlparse==0.4.2
StrEnum==0.4.15
typing_extensions==4.9.0
typing_extensions==4.9.0
urllib3==1.26.7
urllib3==1.26.7
whitenoise
mysqlclient
django-cors-headers

